requirements.txt
# Add project dependencies
openai
# OpenAI Agents SDK (if available via pip, e.g., openai-agents-sdk)
python-dotenv
requests
